-- Bug-fix for beam_fsi_solution (time-accurate fsi). -- Refactored code for flutter solution plots.

+++ src/fluid/conservative_fluid_transient_assembly.cpp
@@ -79,7 +79,7 @@ _linearized_jacobian_solution_product(MAST::ElementBase& elem,
     dynamic_cast<MAST::ConservativeFluidElementBase&>(elem);
     
     const unsigned int
-    n = f.size();
+    n = (unsigned int) f.size();
     
     RealMatrixX
     dummy = RealMatrixX::Zero(n, n);
